+++ src/xpk/core/core.py
@@ -41,7 +41,7 @@ import subprocess
 import sys
 from dataclasses import dataclass
 
-from .. import utils as xpk_utils
+from ..utils import get_user_input, write_tmp_file, xpk_exit, xpk_print
 from .commands import (
     run_command_for_value,
     run_command_with_updates,
@@ -87,8 +87,7 @@ h100_mega_device_type = 'h100-mega-80gb-8'
 
 CAPACITY_TYPE_CONFIG_KEY = 'capacity_type'
 RESERVATION_CONFIG_KEY = 'reservation_id'
-CLUSTER_QUEUE_NAME = 'cluster-queue'
-LOCAL_QUEUE_NAME = 'multislice-queue'
+
 _DEFAULT_POOL_NAME = 'default-pool'
 CLUSTER_RESOURCES_CONFIGMAP = 'resources-configmap'
 CLUSTER_METADATA_CONFIGMAP = 'metadata-configmap'
@@ -114,48 +113,6 @@ class AutoprovisioningConfig:
   maximum_chips: int
 
 
-workload_create_yaml = """apiVersion: jobset.x-k8s.io/v1alpha2
-kind: JobSet
-metadata:
-  name: {args.workload}
-  labels:
-    kueue.x-k8s.io/queue-name: {local_queue_name}  # Name of the LocalQueue
-    xpk.google.com/workload: {args.workload}
-  annotations:
-    alpha.jobset.sigs.k8s.io/exclusive-topology: cloud.google.com/gke-nodepool # 1:1 job replica to node pool assignment
-spec:
-  failurePolicy:
-    maxRestarts: {args.max_restarts}
-  replicatedJobs:
-    - name: slice-job
-      replicas: {args.num_slices}
-      template:
-        spec:
-          parallelism: {system.vms_per_slice}    # Equal to the number of VMs per slice
-          completions: {system.vms_per_slice}    # Same as the above.
-          backoffLimit: 0   # When any pod fails, the job is failed
-          template:
-            metadata:
-              labels:
-                xpk.google.com/workload: {args.workload}
-            spec:
-              schedulerName: {args.scheduler}
-              restartPolicy: Never
-              {affinity}
-              nodeSelector:
-                {accelerator_label}
-                {machine_label}
-                {autoprovisioning_args}
-              priorityClassName: {args.priority}
-              hostNetwork: true
-              dnsPolicy: ClusterFirstWithHostNet
-              terminationGracePeriodSeconds: {args.termination_grace_period_seconds}
-              containers:
-              {container}
-              volumes:
-              {volumes}
-"""
-
 gpu_scheduler_yaml = """schedulerName: {scheduler_name}
               affinity:
                 nodeAffinity:
@@ -174,199 +131,6 @@ gpu_scheduler_yaml = """schedulerName: {scheduler_name}
               """
 
 
-gpu_workload_create_yaml = """apiVersion: jobset.x-k8s.io/v1alpha2
-kind: JobSet
-metadata:
-  name: {args.workload}
-  labels:
-    kueue.x-k8s.io/queue-name: multislice-queue  # Name of the LocalQueue
-    xpk.google.com/workload: {args.workload}
-spec:
-  failurePolicy:
-    maxRestarts: {args.max_restarts}
-  replicatedJobs:
-    - name: slice-job
-      replicas: 1
-      template:
-        spec:
-          parallelism: {args.num_nodes}
-          completions: {args.num_nodes}
-          backoffLimit: 0   # When any pod fails, the job is failed
-          template:
-            metadata:
-              labels:
-                xpk.google.com/workload: {args.workload}
-            spec:
-              {gpu_scheduler}
-              priorityClassName: {args.priority}
-              restartPolicy: Never
-              hostNetwork: true
-              dnsPolicy: ClusterFirstWithHostNet
-              terminationGracePeriodSeconds: {args.termination_grace_period_seconds}
-              tolerations:
-              - operator: "Exists"
-                key: nvidia.com/gpu
-              volumes:
-              {gpu_volume}
-              containers:
-              {gpu_rxdm_image}
-                imagePullPolicy: Always
-                command:
-                - "bash"
-                - "-c"
-                - |
-                  {gpu_rxdm_cmd} &
-                  while [ ! -e "/usr/share/workload/workload_terminated" ]; do sleep 10; echo "sleeping"; done
-                securityContext:
-                  privileged: true
-                volumeMounts:
-                {gpu_tcp_volume}
-                - name: nvidia-install-dir-host
-                  mountPath: /usr/local/nvidia/lib64
-                - name: workload-terminated-volume
-                  mountPath: /usr/share/workload
-                env:
-                - name: LD_LIBRARY_PATH
-                  value: /usr/local/nvidia/lib64
-              {container}
-"""
-
-pw_workload_create_yaml = """apiVersion: jobset.x-k8s.io/v1alpha2
-kind: JobSet
-metadata:
-  name: {args.workload}
-  labels:
-    kueue.x-k8s.io/queue-name: {local_queue_name}  # Name of the LocalQueue
-    xpk.google.com/workload: {args.workload}
-spec:
-  failurePolicy:
-    maxRestarts: {args.max_restarts}
-  successPolicy:
-    operator: "All"
-    targetReplicatedJobs:
-    - {args.targetReplicatedJob}
-  replicatedJobs:
-  - name: worker
-    replicas: {args.num_slices}
-    template:
-      metadata:
-        annotations:
-          alpha.jobset.sigs.k8s.io/exclusive-topology: cloud.google.com/gke-nodepool
-        labels:
-          xpk.google.com/workload: {args.workload}
-      spec:
-        backoffLimit: {backoff_limit}
-        completions: {system.vms_per_slice}
-        parallelism: {system.vms_per_slice}
-        template:
-          spec:
-            terminationGracePeriodSeconds: {args.termination_grace_period_seconds}
-            containers:
-            - args:
-              {pathways_worker_args}
-              image: {args.server_image}
-              imagePullPolicy: Always
-              name: pathways-worker
-              ports:
-              - containerPort: 38677
-              - containerPort: 8471
-              - containerPort: 8080
-              resources:
-                limits:
-                  {resource_type}: {system.chips_per_vm}
-              securityContext:
-                privileged: true
-              volumeMounts:
-              - mountPath: /tmp
-                name: shared-tmp
-            nodeSelector:
-              {accelerator_label}
-              {machine_label}
-              {autoprovisioning_args}
-            priorityClassName: {args.priority}
-            volumes:
-            - hostPath:
-                path: /tmp
-                type: DirectoryOrCreate
-              name: shared-tmp
-  - name: rm
-    replicas: 1
-    template:
-      metadata:
-        labels:
-          xpk.google.com/workload: {args.workload}
-      spec:
-        backoffLimit: 0
-        completions: 1
-        parallelism: 1
-        template:
-          spec:
-            containers:
-            - args:
-              {pathways_rm_args}
-              env:
-              - name: REPLICATED_JOB_NAME
-                valueFrom:
-                  fieldRef:
-                    fieldPath: metadata.annotations['jobset.sigs.k8s.io/replicatedjob-name']
-              - name: JOBSET_NAME
-                valueFrom:
-                  fieldRef:
-                    fieldPath: metadata.annotations['jobset.sigs.k8s.io/jobset-name']
-              - name: HOST_ADDRESS
-                value: $(JOBSET_NAME)-$(REPLICATED_JOB_NAME)-0-0.$(JOBSET_NAME)
-              - name: TPU_SKIP_MDS_QUERY
-                value: "true"
-              image: {args.server_image}
-              imagePullPolicy: Always
-              name: pathways-rm
-              ports:
-              - containerPort: 38677
-              resources:
-                limits:
-                  cpu: "4"
-                  memory: 8G
-              securityContext:
-                privileged: true
-              volumeMounts:
-              - mountPath: /tmp
-                name: shared-tmp
-            nodeSelector:
-              cloud.google.com/gke-nodepool: cpu-rm-np
-            volumes:
-            - hostPath:
-                path: /tmp
-                type: DirectoryOrCreate
-              name: shared-tmp
-  - name: proxy
-    replicas: 1
-    template:
-      metadata:
-        labels:
-          xpk.google.com/workload: {args.workload}
-      spec:
-        backoffLimit: 0
-        completions: 1
-        parallelism: 1
-        template:
-          spec:
-            containers:
-            - args:
-              {pathways_proxy_args}
-              image: {args.proxy_server_image}
-              imagePullPolicy: Always
-              name: pathways-proxy
-              ports:
-              - containerPort: 38676
-              resources:
-                limits:
-                  cpu: "24"
-                  memory: 100G
-            nodeSelector:
-              cloud.google.com/gke-nodepool: cpu-proxy-np
-  {user_workload}
-"""
-
 script_dir_dockerfile = """FROM {base_docker_image}
 
 # Set the working directory in the container
@@ -378,111 +142,6 @@ COPY . .
 WORKDIR /app
 """
 
-cluster_set_crd_yaml = """apiVersion: kueue.x-k8s.io/v1beta1
-kind: ResourceFlavor
-metadata:
-  name: {cluster_hardware_name}
-spec:
-  nodeLabels:
-    {accelerator_label}
-    {machine_label}
----
-{pw_resource_flavors}
-apiVersion: kueue.x-k8s.io/v1beta1
-kind: ClusterQueue
-metadata:
-  name: {cluster_queue_name}
-spec:
-  preemption:
-      reclaimWithinCohort: Never # Don't preempt other queues in the cohort.
-      withinClusterQueue: LowerPriority
-  namespaceSelector: {{}} # match all.
-  resourceGroups:
-  {covered_resources_config}
-  {pw_resources_kueue}
----
-apiVersion: kueue.x-k8s.io/v1beta1
-kind: LocalQueue
-metadata:
-  namespace: default
-  name: {local_queue_name}
-spec:
-  clusterQueue: {cluster_queue_name}
----
-apiVersion: scheduling.k8s.io/v1
-kind: PriorityClass
-metadata:
-  name: very-low
-value: 100
-globalDefault: false
-description: "Very Low"
----
-apiVersion: scheduling.k8s.io/v1
-kind: PriorityClass
-metadata:
-  name: low
-value: 250
-globalDefault: false
-description: "Low"
----
-apiVersion: scheduling.k8s.io/v1
-kind: PriorityClass
-metadata:
-  name: medium
-value: 500
-globalDefault: false
-description: "Medium"
----
-apiVersion: scheduling.k8s.io/v1
-kind: PriorityClass
-metadata:
-  name: high
-value: 750
-globalDefault: false
-description: "High"
----
-apiVersion: scheduling.k8s.io/v1
-kind: PriorityClass
-metadata:
-  name: very-high
-value: 1000
-globalDefault: false
-description: "Very High"
-"""
-
-cluster_preheat_yml = """
-apiVersion: apps/v1
-kind: DaemonSet
-metadata:
-  name: {cachekey}
-  labels:
-    k8s-app: {cachekey}
-spec:
-  selector:
-    matchLabels:
-      k8s-app: {cachekey}
-  updateStrategy:
-    type: RollingUpdate
-  template:
-    metadata:
-      labels:
-        name: {cachekey}
-        k8s-app: {cachekey}
-    spec:
-      affinity:
-        nodeAffinity:
-          requiredDuringSchedulingIgnoredDuringExecution:
-            nodeSelectorTerms:
-            - matchExpressions:
-              - key: {nodeSelectorKey}
-                operator: Exists
-      tolerations:
-      - operator: "Exists"
-      containers:
-      - image: {image_name}
-        name: {cachekey}
-        command: [ "sleep", "inf" ]
-"""
 
 cluster_configmap_yaml = """kind: ConfigMap
 apiVersion: v1
@@ -576,14 +235,6 @@ spec:
 """
 
 
-PathwaysExpectedInstancesMap = {
-    'v5p': 'v5',
-    'v5litepod': 'v5e',
-    'v4': 'v4',
-    'v3': 'v3',
-}
-
-
 def add_zone_and_project(args):
   """Obtains the zone and project names from gcloud configs if not defined.
 
@@ -594,7 +245,7 @@ def add_zone_and_project(args):
     args.project = get_project()
   if not args.zone:
     args.zone = get_zone()
-  xpk_utils.xpk_print(f'Working on {args.project=} and {args.zone}')
+  xpk_print(f'Working on {args.project=} and {args.zone}')
 
 
 def parse_env_config(args, tensorboard_config, system: SystemCharacteristics):
@@ -750,16 +401,12 @@ def update_gke_cluster_with_clouddns(args) -> int:
       f' --cluster-dns-domain={args.cluster}-domain'
       ' --quiet'
   )
-  xpk_utils.xpk_print(
-      'Updating GKE cluster to use Cloud DNS, may take a while!'
-  )
+  xpk_print('Updating GKE cluster to use Cloud DNS, may take a while!')
   return_code = run_command_with_updates(
       command, 'GKE Cluster Update to enable Cloud DNS', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
-        f'GKE Cluster Update request returned ERROR {return_code}'
-    )
+    xpk_print(f'GKE Cluster Update request returned ERROR {return_code}')
     return 1
   return 0
 
@@ -782,16 +429,14 @@ def upgrade_gke_control_plane_version(args, default_rapid_gke_version) -> int:
       ' --master'
       ' --quiet'
   )
-  xpk_utils.xpk_print(
-      "Updating GKE cluster's control plane version, may take a while!"
-  )
+  xpk_print("Updating GKE cluster's control plane version, may take a while!")
   return_code = run_command_with_updates(
       command,
       'GKE Cluster control plane version update to enable Cloud DNS',
       args,
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         "GKE cluster's control plane version update request returned"
         f' ERROR {return_code}'
     )
@@ -811,7 +456,7 @@ def upgrade_gke_nodepools_version(args, default_rapid_gke_version) -> int:
   """
   existing_node_pool_names, return_code = get_all_nodepools_programmatic(args)
   if return_code != 0:
-    xpk_utils.xpk_print('Listing all node pools failed!')
+    xpk_print('Listing all node pools failed!')
     return return_code
 
   # Batch execution to upgrade node pools simultaneously
@@ -829,14 +474,12 @@ def upgrade_gke_nodepools_version(args, default_rapid_gke_version) -> int:
     task_names.append(f'Upgrading node pool {node_pool_name}.')
 
   for i, command in enumerate(commands):
-    xpk_utils.xpk_print(
-        f'To complete {task_names[i]} we are executing {command}'
-    )
+    xpk_print(f'To complete {task_names[i]} we are executing {command}')
   max_return_code = run_commands(
       commands, 'Update GKE node pools to default RAPID GKE version', task_names
   )
   if max_return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         'GKE node pools update to default RAPID GKE version returned ERROR:'
         f' {max_return_code}'
     )
@@ -882,7 +525,7 @@ def create_cluster_network(args, index) -> int:
   """
   existing_network_names, return_code = get_all_networks_programmatic(args)
   if return_code > 0:
-    xpk_utils.xpk_print('Listing all networks failed!')
+    xpk_print('Listing all networks failed!')
     return return_code
 
   network_name = f'{args.cluster}-net-{index}'
@@ -897,12 +540,10 @@ def create_cluster_network(args, index) -> int:
     )
 
     if return_code != 0:
-      xpk_utils.xpk_print(
-          f'Create Cluster Network request returned ERROR {return_code}'
-      )
+      xpk_print(f'Create Cluster Network request returned ERROR {return_code}')
       return 1
   else:
-    xpk_utils.xpk_print(f'Reusing existing network {network_name}')
+    xpk_print(f'Reusing existing network {network_name}')
 
   return 0
 
@@ -919,7 +560,7 @@ def create_cluster_subnet(args, index) -> int:
   """
   existing_subnet_names, return_code = get_all_subnets_programmatic(args)
   if return_code > 0:
-    xpk_utils.xpk_print('Listing all subnets failed!')
+    xpk_print('Listing all subnets failed!')
     return return_code
   subnet_name = f'{args.cluster}-{zone_to_region(args.zone)}-sub-{index}'
   if subnet_name not in existing_subnet_names:
@@ -934,12 +575,10 @@ def create_cluster_subnet(args, index) -> int:
     )
 
     if return_code != 0:
-      xpk_utils.xpk_print(
-          f'Create Cluster Subnet request returned ERROR {return_code}'
-      )
+      xpk_print(f'Create Cluster Subnet request returned ERROR {return_code}')
       return 1
   else:
-    xpk_utils.xpk_print(f'Reusing existing subnet {subnet_name}')
+    xpk_print(f'Reusing existing subnet {subnet_name}')
 
   return 0
 
@@ -955,7 +594,7 @@ def delete_cluster_subnets(args) -> int:
   """
   existing_subnet_names, return_code = get_all_subnets_programmatic(args)
   if return_code > 0:
-    xpk_utils.xpk_print('Listing all subnets failed!')
+    xpk_print('Listing all subnets failed!')
     return return_code
 
   for subnet_name in existing_subnet_names:
@@ -969,12 +608,10 @@ def delete_cluster_subnets(args) -> int:
     )
 
     if return_code != 0:
-      xpk_utils.xpk_print(
-          f'Delete Cluster Subnet request returned ERROR {return_code}'
-      )
+      xpk_print(f'Delete Cluster Subnet request returned ERROR {return_code}')
       return 1
     else:
-      xpk_utils.xpk_print(f'Deleted existing subnet {subnet_name}')
+      xpk_print(f'Deleted existing subnet {subnet_name}')
 
   return 0
 
@@ -993,7 +630,7 @@ def create_cluster_firewall_rule(args, index) -> int:
       get_all_firewall_rules_programmatic(args)
   )
   if return_code > 0:
-    xpk_utils.xpk_print('Listing all firewall rules failed!')
+    xpk_print('Listing all firewall rules failed!')
     return return_code
   firewall_rule_name = f'{args.cluster}-internal-{index}'
   if firewall_rule_name not in existing_firewall_rules_names:
@@ -1007,12 +644,12 @@ def create_cluster_firewall_rule(args, index) -> int:
     )
 
     if return_code != 0:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'Create Cluster Firewall Rule request returned ERROR {return_code}'
       )
       return 1
   else:
-    xpk_utils.xpk_print(f'Reusing existing firewall rule {firewall_rule_name}')
+    xpk_print(f'Reusing existing firewall rule {firewall_rule_name}')
   return 0
 
 
@@ -1026,14 +663,14 @@ def create_cluster_network_config(args) -> int:
     0 if successful and 1 otherwise.
   """
   yml_string = cluster_network_yaml.format(cluster_name=args.cluster)
-  tmp = xpk_utils.write_tmp_file(yml_string)
+  tmp = write_tmp_file(yml_string)
   command = f'kubectl apply -f {str(tmp.file.name)}'
 
   return_code = run_command_with_updates(
       command, 'GKE Cluster Create Network Config', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'GKE Cluster Create ConfigMap request returned ERROR {return_code}'
     )
     return 1
@@ -1055,7 +692,7 @@ def print_reservations(args) -> int:
       command, 'Get all reservations in the project', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(f'Get all reservations returned ERROR {return_code}')
+    xpk_print(f'Get all reservations returned ERROR {return_code}')
     return 1
   return 0
 
@@ -1075,8 +712,8 @@ def verify_reservation_exists(args) -> int:
   )
   return_code = run_command_with_updates(command, 'Describe reservation', args)
   if return_code != 0:
-    xpk_utils.xpk_print(f'Describe reservation returned ERROR {return_code}')
-    xpk_utils.xpk_print('Please confirm that your reservation name is correct.')
+    xpk_print(f'Describe reservation returned ERROR {return_code}')
+    xpk_print('Please confirm that your reservation name is correct.')
     return 1
   return 0
 
@@ -1113,7 +750,7 @@ def get_capacity_type(args) -> tuple[CapacityType, int]:
   if num_types == 0:
     capacity_type = CapacityType.UNKNOWN
   elif num_types != 1:
-    xpk_utils.xpk_print(
+    xpk_print(
         'ERROR: User specified more than one of the following arguments. Please'
         ' specify only one of `--reservation=$RESERVATION_NAME`, `--on-demand`'
         ' or `--spot`.'
@@ -1149,7 +786,7 @@ def get_capacity_arguments_from_capacity_type(
           f'--reservation-affinity=specific --reservation={args.reservation}'
       )
     case _:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'Unknown capacity type: {capacity_type}. Unable to determine'
           ' capacity args.'
       )
@@ -1181,7 +818,7 @@ def get_capacity_node_selectors_from_capacity_type(
     case CapacityType.RESERVATION.name:
       node_selector = f'cloud.google.com/reservation-name: {args.reservation}'
     case _:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'Unknown capacity type: {capacity_type}. Unable to determine the'
           ' node selectors.'
       )
@@ -1200,7 +837,7 @@ def create_or_update_cluster_configmap(configmap_yml: dict) -> int:
   commands = []
   task_names = []
   for configmap_name, yml_string in configmap_yml.items():
-    tmp = xpk_utils.write_tmp_file(yml_string)
+    tmp = write_tmp_file(yml_string)
     command = f'kubectl apply -f {str(tmp.file.name)}'
     commands.append(command)
     task_name = f'ConfigMap CreateOrUpdate-{configmap_name}'
@@ -1210,7 +847,7 @@ def create_or_update_cluster_configmap(configmap_yml: dict) -> int:
       commands, 'GKE Cluster CreateOrUpdate ConfigMap(s)', task_names
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         'GKE Cluster Create/Update ConfigMap(s) request returned ERROR'
         f' {return_code}'
     )
@@ -1277,7 +914,7 @@ def create_cluster_configmaps(
   # Capacity Type.
   capacity_type, return_code = get_capacity_type(args)
   if return_code != 0:
-    xpk_utils.xpk_print('Unable to determine capacity type.')
+    xpk_print('Unable to determine capacity type.')
     return return_code
   metadata += f'\n  {CAPACITY_TYPE_CONFIG_KEY}: {capacity_type.name}'
   # Reservation ID if applicable.
@@ -1310,9 +947,7 @@ def get_cluster_configmap(args, configmap_name) -> dict[str, str] | None:
       command, 'GKE Cluster Get ConfigMap', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
-        f'GKE Cluster Get ConfigMap request returned ERROR {return_code}'
-    )
+    xpk_print(f'GKE Cluster Get ConfigMap request returned ERROR {return_code}')
     return None
 
   config_map = {}
@@ -1350,7 +985,7 @@ def create_vertex_tensorboard(args) -> dict:
       tensorboard_name=tensorboard_name,
   )
   if instance_id:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Tensorboard instance {tensorboard_name} is successfully created.'
     )
     tensorboard_config['tensorboard_region'] = args.tensorboard_region
@@ -1374,7 +1009,7 @@ def create_vertex_experiment(args) -> dict:
   cluster_config_map = get_cluster_configmap(args, metadata_configmap_name)
 
   if cluster_config_map is None or 'tensorboard_name' not in cluster_config_map:
-    xpk_utils.xpk_print(
+    xpk_print(
         'No Vertex Tensorboard instance has been created in cluster create. Run'
         ' `xpk cluster create --create-vertex-tensorboard` before running `xpk'
         ' workload create --use-vertex-tensorboard` to create a Vertex'
@@ -1406,7 +1041,7 @@ def create_vertex_experiment(args) -> dict:
   if tensorboard_url is None:
     return None
 
-  xpk_utils.xpk_print(f'You can view Vertex Tensorboard at: {tensorboard_url}')
+  xpk_print(f'You can view Vertex Tensorboard at: {tensorboard_url}')
   return tensorboard_config
 
 
@@ -1428,7 +1063,7 @@ def get_all_clusters_programmatic(args) -> tuple[list[str], int]:
       command, 'Find if Cluster Exists', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(f'Find if Cluster Exists returned ERROR {return_code}')
+    xpk_print(f'Find if Cluster Exists returned ERROR {return_code}')
     return [], return_code
 
   return raw_cluster_output.splitlines(), 0
@@ -1453,12 +1088,10 @@ def is_cluster_using_clouddns(args) -> bool:
       args,
   )
   if return_code != 0:
-    xpk_utils.xpk_exit(return_code)
+    xpk_exit(return_code)
   cloud_dns_matches = int(cloud_dns_matches)
   if cloud_dns_matches > 0:
-    xpk_utils.xpk_print(
-        'Cloud DNS is enabled on the cluster, no update needed.'
-    )
+    xpk_print('Cloud DNS is enabled on the cluster, no update needed.')
     return True
   return False
 
@@ -1474,7 +1107,7 @@ def update_cluster_with_clouddns_if_necessary(args) -> int:
   """
   all_clusters, return_code = get_all_clusters_programmatic(args)
   if return_code > 0:
-    xpk_utils.xpk_print('Listing all clusters failed!')
+    xpk_print('Listing all clusters failed!')
     return 1
   if args.cluster in all_clusters:
     # If cluster is already using clouddns, no update necessary!
@@ -1482,20 +1115,18 @@ def update_cluster_with_clouddns_if_necessary(args) -> int:
       return 0
     cluster_update_return_code = update_gke_cluster_with_clouddns(args)
     if cluster_update_return_code > 0:
-      xpk_utils.xpk_print('Updating GKE cluster to use CloudDNS failed!')
+      xpk_print('Updating GKE cluster to use CloudDNS failed!')
       return cluster_update_return_code
 
     # Find default rapid control plane version and update the control plane to the same.
     server_config_return_code, gke_server_config = get_gke_server_config(args)
     if server_config_return_code != 0:
-      xpk_utils.xpk_exit(server_config_return_code)
+      xpk_exit(server_config_return_code)
     upgrade_master_return_code = upgrade_gke_control_plane_version(
         args, gke_server_config.default_rapid_gke_version
     )
     if upgrade_master_return_code > 0:
-      xpk_utils.xpk_print(
-          "Updating GKE cluster's control plane upgrade failed!"
-      )
+      xpk_print("Updating GKE cluster's control plane upgrade failed!")
       return upgrade_master_return_code
 
     # Upgrade nodepools version after the master upgrade.
@@ -1503,7 +1134,7 @@ def update_cluster_with_clouddns_if_necessary(args) -> int:
         args, gke_server_config.default_rapid_gke_version
     )
     if node_pool_update_code > 0:
-      xpk_utils.xpk_print('Upgrading nodepools version failed!')
+      xpk_print('Upgrading nodepools version failed!')
       return node_pool_update_code
   return 0
 
@@ -1529,7 +1160,7 @@ def get_nodepool_zone(args, nodepool_name) -> tuple[int, str]:
       command, 'Get Node Pool Zone', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(f'Get Node Pool Zone returned ERROR {return_code}')
+    xpk_print(f'Get Node Pool Zone returned ERROR {return_code}')
     return 1, None
 
   return 0, nodepool_zone.strip()
@@ -1551,7 +1182,7 @@ def check_cluster_resources(args, system) -> tuple[bool, bool]:
   resources_configmap_name = f'{args.cluster}-{CLUSTER_RESOURCES_CONFIGMAP}'
   resources_config_map = get_cluster_configmap(args, resources_configmap_name)
   if resources_config_map is None:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'No ConfigMap exist for cluster with the name {resources_config_map}.'
         ' Cluster resources check will be skipped.'
     )
@@ -1582,7 +1213,7 @@ def get_all_nodepools_programmatic(args) -> tuple[list[str], int]:
       command, 'Get All Node Pools', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(f'Get All Node Pools returned ERROR {return_code}')
+    xpk_print(f'Get All Node Pools returned ERROR {return_code}')
     return [], 1
 
   return raw_nodepool_output.splitlines(), 0
@@ -1602,7 +1233,7 @@ def get_all_networks_programmatic(args) -> tuple[list[str], int]:
       command, 'Get All Networks', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(f'Get All Networks returned ERROR {return_code}')
+    xpk_print(f'Get All Networks returned ERROR {return_code}')
     return [], 1
 
   return raw_network_output.splitlines(), 0
@@ -1627,7 +1258,7 @@ def get_all_subnets_programmatic(args) -> tuple[list[str], int]:
       command, 'Get All Subnets', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(f'Get All Subnets returned ERROR {return_code}')
+    xpk_print(f'Get All Subnets returned ERROR {return_code}')
     return [], 1
 
   all_outputs = raw_subnets_output.splitlines()
@@ -1653,7 +1284,7 @@ def get_all_firewall_rules_programmatic(args) -> tuple[list[str], int]:
       command, 'Get All Firewall Rules', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(f'Get All Firewall Rules returned ERROR {return_code}')
+    xpk_print(f'Get All Firewall Rules returned ERROR {return_code}')
     return [], 1
 
   return raw_subnets_output.splitlines(), 0
@@ -1711,44 +1342,44 @@ def run_gke_node_pool_create_command(
     0 if successful and 1 otherwise.
   """
   device_type = args.tpu_type if args.tpu_type else args.device_type
-  xpk_utils.xpk_print(
+  xpk_print(
       f'Creating {args.num_slices} node pool or pools of {device_type}\n'
       f'We assume that the underlying system is: {system}'
   )
   existing_node_pool_names, return_code = get_all_nodepools_programmatic(args)
   if return_code > 0:
-    xpk_utils.xpk_print('Listing all node pools failed!')
+    xpk_print('Listing all node pools failed!')
     return return_code
 
   capacity_type, return_code = get_capacity_type(args)
   if return_code > 0:
-    xpk_utils.xpk_print('Parsing capacity type failed!')
+    xpk_print('Parsing capacity type failed!')
     return return_code
   if capacity_type == CapacityType.UNKNOWN:
     return_code = print_reservations(args)
-    xpk_utils.xpk_print(
+    xpk_print(
         'ERROR: User needs to provide the capacity type. Please specify one of'
         ' the following `--reservation=$RESERVATION_NAME`, `--on-demand`'
         ' or `--spot`. See the above list of reservations to choose from.'
     )
     if return_code > 0:
-      xpk_utils.xpk_print('Listing all reservations failed!')
+      xpk_print('Listing all reservations failed!')
     return_code = 1
   capacity_args, return_code = get_capacity_arguments_from_capacity_type(
       args, capacity_type
   )
   if return_code > 0:
-    xpk_utils.xpk_print('Parsing capacity arguments failed!')
+    xpk_print('Parsing capacity arguments failed!')
     return return_code
 
   if system.accelerator_type == AcceleratorType['GPU']:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Creating 1 node pool with {args.num_nodes} nodes of'
         f' {system.device_type}\nUnderlyingly, we assume that means: {system}'
     )
     desired_node_pool_names = [f'{args.cluster}-np-0']
   else:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Creating {args.num_slices} node pool or pools of'
         f' {system.device_type}\nUnderlyingly, we assume that means: {system}'
     )
@@ -1767,7 +1398,7 @@ def run_gke_node_pool_create_command(
       return 1
 
     if existing_node_pool_zone and existing_node_pool_zone != args.zone:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'Cluster {args.cluster} already has nodepools in zone:'
           f' {existing_node_pool_zone}. Use the same zone to update nodepools'
           ' in the cluster.'
@@ -1800,20 +1431,20 @@ def run_gke_node_pool_create_command(
   if delete_commands:
     will_delete = True
     if node_pools_to_delete and not args.force:
-      will_delete = xpk_utils.get_user_input(
+      will_delete = get_user_input(
           f'Planning to delete {len(node_pools_to_delete)} node pools including'
           f' {node_pools_to_delete}. \nDo you wish to delete: y (yes) / n'
           ' (no):\n'
       )
     if not will_delete:
-      xpk_utils.xpk_print(
+      xpk_print(
           'You have requested to not delete the existing nodepools in the'
           ' cluster. There will be no change to the cluster.'
       )
       return 1
 
     for i, command in enumerate(delete_commands):
-      xpk_utils.xpk_print(
+      xpk_print(
           f'To complete {delete_task_names[i]} we are executing {command}'
       )
     max_return_code = run_commands(
@@ -1823,7 +1454,7 @@ def run_gke_node_pool_create_command(
         dry_run=args.dry_run,
     )
     if max_return_code != 0:
-      xpk_utils.xpk_print(f'Delete Nodepools returned ERROR {max_return_code}')
+      xpk_print(f'Delete Nodepools returned ERROR {max_return_code}')
       return 1
 
     # Update {args.cluster}-{_CLUSTER_RESOURCES_CONFIGMAP} ConfigMap to 'y': '0'
@@ -1931,9 +1562,7 @@ def run_gke_node_pool_create_command(
       create_task_names.append(task)
 
   for i, command in enumerate(create_commands):
-    xpk_utils.xpk_print(
-        f'To complete {create_task_names[i]} we are executing {command}'
-    )
+    xpk_print(f'To complete {create_task_names[i]} we are executing {command}')
   max_return_code = run_commands(
       create_commands,
       'Create Nodepools',
@@ -1941,207 +1570,11 @@ def run_gke_node_pool_create_command(
       dry_run=args.dry_run,
   )
   if max_return_code != 0:
-    xpk_utils.xpk_print(f'Create Nodepools returned ERROR {max_return_code}')
+    xpk_print(f'Create Nodepools returned ERROR {max_return_code}')
     return 1
 
-  xpk_utils.xpk_print('Create or delete node pool request complete.')
+  xpk_print('Create or delete node pool request complete.')
   return 0
-
-
-def install_kueue_on_cluster(args) -> int:
-  """Install Kueue on the cluster.
-
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    0 if successful and 1 otherwise.
-  """
-  command = (
-      'kubectl apply --server-side --force-conflicts -f'
-      ' https://github.com/kubernetes-sigs/kueue/releases/download/v0.6.1/manifests.yaml'
-  )
-  task = 'Set Kueue On Cluster'
-  return_code = run_command_with_updates_retry(command, task, args)
-  if return_code != 0:
-    xpk_utils.xpk_print(f'{task} returned ERROR {return_code}')
-  return return_code
-
-
-def enable_kueue_credentials(
-    args,
-    system: SystemCharacteristics,
-    autoprovisioning_config: AutoprovisioningConfig | None,
-) -> int:
-  """Enable Kueue credentials.
-
-  Args:
-    args: user provided arguments for running the command.
-    system: system level arguments.
-    autoprovisioning_config: Autoprovisioning config to configure kueue with if
-        autoprovisioning is enabled.
-
-  Returns:
-    0 if successful and 1 otherwise.
-  """
-  device_type = system.device_type
-  cluster_hardware_name = f'{args.num_slices}x{device_type}'
-  resource_type = AcceleratorTypeToAcceleratorCharacteristics[
-      system.accelerator_type
-  ].resource_type
-
-  autoprovisioning_enabled = False
-  if autoprovisioning_config:
-    # Determine total resources available based on autoprovisioning max chips.
-    autoprovisioning_enabled = True
-    total_chips = autoprovisioning_config.maximum_chips
-    cluster_hardware_name = f'{system.gke_accelerator}'
-  else:
-    # Determine total chips based on user specified topology.
-    total_chips = get_total_chips_requested_from_args(args, system)
-
-  covered_resources_config = get_kueue_covered_resources_config(
-      cluster_hardware_name=cluster_hardware_name,
-      resource_type=resource_type,
-      total_chips=total_chips,
-  )
-  yml_string = cluster_set_crd_yaml.format(
-      system=system,
-      cluster_hardware_name=cluster_hardware_name,
-      accelerator_label=create_accelerator_label(
-          system.accelerator_type, system
-      ),
-      machine_label=create_machine_label(
-          system.accelerator_type, system, autoprovisioning_enabled
-      ),
-      covered_resources_config=covered_resources_config,
-      resource_type=AcceleratorTypeToAcceleratorCharacteristics[
-          system.accelerator_type
-      ].resource_type,
-      pw_resource_flavors=add_pw_resource_flavors(args),
-      pw_resources_kueue=add_pw_resources_to_kueue(args),
-      cluster_queue_name=CLUSTER_QUEUE_NAME,
-      local_queue_name=LOCAL_QUEUE_NAME,
-  )
-
-  tmp = xpk_utils.write_tmp_file(yml_string)
-  command = f'kubectl apply -f {str(tmp.file.name)}'
-  # For kueue setup, we see a timeout error due to the webhook not
-  # being ready. Let's retry and wait a few seconds.
-  task = 'Applying Kueue Credentials'
-  retry_attempts = 3
-  return_code = run_command_with_updates_retry(
-      command, task, args, num_retry_attempts=retry_attempts
-  )
-  if return_code != 0:
-    # We have seen some scenarios where credentials need a few minutes for kueue
-    # and jobset installation to be ready before credentials can be applied.
-    # As a workaround we will retry again with longer wait times.
-    retry_wait_seconds = 60
-    xpk_utils.xpk_print(
-        f'{task} still not successful. Retrying {retry_attempts} more timeswith'
-        f' increased wait time of {retry_wait_seconds} seconds between tries.'
-        ' Kueue Credentials need Kueue system to be ready which can take some'
-        ' time.'
-    )
-    return_code = run_command_with_updates_retry(
-        command=command,
-        task=task,
-        args=args,
-        num_retry_attempts=retry_attempts,
-        wait_seconds=retry_wait_seconds,
-    )
-    if return_code != 0:
-      xpk_utils.xpk_print(f'{task} returned ERROR {return_code}')
-  return return_code
-
-
-# TODO(roshanin): Organize Pathways helpers in another file.
-def add_pw_resource_flavors(args):
-  """Add resource flavors required for Pathways enabled clusters."""
-  resource_flavor_yaml = """apiVersion: kueue.x-k8s.io/v1beta1
-kind: ResourceFlavor
-metadata:
-  name: cpu-rm
-spec:
-  nodeLabels:
-    cloud.google.com/gke-nodepool: cpu-rm-np
----
-apiVersion: kueue.x-k8s.io/v1beta1
-kind: ResourceFlavor
-metadata:
-  name: cpu-proxy
-spec:
-  nodeLabels:
-    cloud.google.com/gke-nodepool: cpu-proxy-np
----
-apiVersion: kueue.x-k8s.io/v1beta1
-kind: ResourceFlavor
-metadata:
-  name: cpu-user
-spec:
-  nodeLabels:
-    cloud.google.com/gke-nodepool: cpu-user-np
----"""
-  if args.enable_pathways:
-    return resource_flavor_yaml
-  return ''
-
-
-def add_pw_resources_to_kueue(args):
-  """Add resource flavors required for Pathways, to the cluster queue."""
-  resources_yaml = """- coveredResources: ["cpu", "memory"]
-    flavors:
-    - name: cpu-rm
-      resources:
-      - name: "cpu"
-        nominalQuota: 80
-      - name: "memory"
-        nominalQuota: 160G
-    - name: cpu-proxy
-      resources:
-      - name: "cpu"
-        nominalQuota: 480
-      - name: "memory"
-        nominalQuota: 2000G
-    - name: cpu-user
-      resources:
-      - name: "cpu"
-        nominalQuota: 480
-      - name: "memory"
-        nominalQuota: 2000G"""
-  if args.enable_pathways:
-    return resources_yaml
-  return ''
-
-
-def get_kueue_covered_resources_config(
-    cluster_hardware_name, resource_type, total_chips
-) -> str:
-  """Gets Kueue covered resources configuration.
-
-  Args:
-    cluster_hardware_name: cluster hardware name.
-    resource_type: resource type of tpu or gpu.
-    total_chips: total number of chips for the specific resource type.
-
-  Returns:
-    A string of Kueue covered resources configuration.
-  """
-  config_format = """
-  - coveredResources: ["{resource_type}"]
-    flavors:
-    - name: {cluster_hardware_name}
-      resources:
-      - name: "{resource_type}"
-        nominalQuota: {total_chips}
-  """
-  config_string = config_format.format(
-      cluster_hardware_name=cluster_hardware_name,
-      resource_type=resource_type,
-      total_chips=total_chips,
-  )
-  return config_string
 
 
 # TODO(vbarr): Remove this function when jobsets gets enabled by default on
@@ -2163,8 +1596,8 @@ def set_jobset_on_cluster(args) -> int:
   return_code = run_command_with_updates_retry(command, task, args)
 
   if return_code != 0:
-    xpk_utils.xpk_print(f'{task} returned with ERROR {return_code}.\n')
-    xpk_utils.xpk_print(
+    xpk_print(f'{task} returned with ERROR {return_code}.\n')
+    xpk_print(
         "This LIKELY means you're missing Kubernetes Permissions, you can"
         ' validate this by checking if the error references permission problems'
         ' such as `requires one of ["container.*"] permission(s)`. Follow our'
@@ -2203,7 +1636,7 @@ def install_nccl_on_cluster(args, system: SystemCharacteristics) -> int:
   )
 
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Install NCCL Plugin On Cluster request returned ERROR {return_code}'
     )
     return 1
@@ -2266,9 +1699,7 @@ def get_gke_server_config(args) -> tuple[int, GkeServerConfig | None]:
         hide_error=True,
     )
     if return_code != 0:
-      xpk_utils.xpk_print(
-          f'Unable to get server config for {command_description}.'
-      )
+      xpk_print(f'Unable to get server config for {command_description}.')
       return return_code, None
     command_outputs.append(cmd_output)
 
@@ -2302,16 +1733,16 @@ def get_gke_control_plane_version(
   is_valid_version = master_gke_version in gke_server_config.valid_versions
 
   if not is_valid_version:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Planned GKE Version: {master_gke_version}\n Valid Versions:'
         f'\n{gke_server_config.valid_versions}\nRecommended / Default GKE'
         f' Version: {gke_server_config.default_rapid_gke_version}'
     )
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Error: Planned GKE Version {master_gke_version} is not valid.'
         f'Checks failed: Is Version Valid: {is_valid_version}'
     )
-    xpk_utils.xpk_print(
+    xpk_print(
         'Please select a gke version from the above list using --gke-version=x'
         ' argument or rely on the default gke version:'
         f' {gke_server_config.default_rapid_gke_version}'
@@ -2348,7 +1779,7 @@ def get_gke_node_pool_version(
       command, command_description, args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Unable to get server config for command: {command_description}.'
     )
     return return_code, None
@@ -2365,7 +1796,7 @@ def get_gke_node_pool_version(
   # In rare cases, user's provided gke version may be invalid, but gke will return an error if so.
   # An example scenario is if the user provided gke version is greater than the master version.
   if not is_supported_node_pool_version:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Planned node pool version {node_pool_gke_version} is not supported in'
         ' valid version'
         f' {gke_server_config.valid_versions}\nPlease adjust the gke version'
@@ -2399,7 +1830,7 @@ def validate_docker_image(docker_image, args) -> int:
       command, 'Validate Docker Image', args, verbose=False
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         'Failed to validate your docker image, check that the docker image'
         f' exists. You may be able to find the {docker_image} in {project}.'
         ' If the docker image exists, the service account of this'
@@ -2429,11 +1860,11 @@ def build_docker_image_from_base_image(args, verbose=True) -> tuple[int, str]:
   docker_file = script_dir_dockerfile.format(
       base_docker_image=args.base_docker_image,
   )
-  tmp = xpk_utils.write_tmp_file(docker_file)
+  tmp = write_tmp_file(docker_file)
   docker_build_command = (
       f'docker build -f {str(tmp.file.name)} -t {docker_name} {args.script_dir}'
   )
-  xpk_utils.xpk_print(f'Building {args.script_dir} into docker image.')
+  xpk_print(f'Building {args.script_dir} into docker image.')
   return_code = run_command_with_updates(
       docker_build_command,
       'Building script_dir into docker image',
@@ -2441,12 +1872,12 @@ def build_docker_image_from_base_image(args, verbose=True) -> tuple[int, str]:
       verbose=verbose,
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         'Failed to add script_dir to docker image, check the base docker image.'
         f' You should be able to navigate to the URL {args.base_docker_image}'
         f' in {args.project}.'
     )
-    xpk_utils.xpk_exit(1)
+    xpk_exit(1)
 
   # Pick a randomly generated `tag_length` character docker tag.
   tag_length = 4
@@ -2456,9 +1887,7 @@ def build_docker_image_from_base_image(args, verbose=True) -> tuple[int, str]:
   tag_datetime = datetime.datetime.now().strftime('%Y-%m-%d-%H-%M-%S')
   tag_name = f'{tag_random_prefix}-{tag_datetime}'
   cloud_docker_image = f'gcr.io/{args.project}/{docker_name}:{tag_name}'
-  xpk_utils.xpk_print(
-      f'Adding Docker Image: {cloud_docker_image} to {args.project}'
-  )
+  xpk_print(f'Adding Docker Image: {cloud_docker_image} to {args.project}')
 
   # Tag the docker image.
   tag_docker_image_command = f'docker tag {docker_name} {cloud_docker_image}'
@@ -2466,12 +1895,12 @@ def build_docker_image_from_base_image(args, verbose=True) -> tuple[int, str]:
       tag_docker_image_command, 'Tag Docker Image', args, verbose=verbose
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Failed to tag docker image with tag: {tag_name}.'
         f' You should be able to navigate to the URL {cloud_docker_image} in'
         f' {args.project}.'
     )
-    xpk_utils.xpk_exit(1)
+    xpk_exit(1)
 
   # Upload image to Artifact Registry.
   upload_docker_image_command = f'docker push {cloud_docker_image}'
@@ -2479,12 +1908,12 @@ def build_docker_image_from_base_image(args, verbose=True) -> tuple[int, str]:
       upload_docker_image_command, 'Upload Docker Image', args, verbose=verbose
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         'Failed to upload docker image.'
         f' You should be able to navigate to the URL {cloud_docker_image} in'
         f' {args.project}.'
     )
-    xpk_utils.xpk_exit(1)
+    xpk_exit(1)
   return return_code, cloud_docker_image
 
 
@@ -2509,8 +1938,8 @@ def check_if_workload_exists(args) -> bool:
   )
 
   if return_code != 0:
-    xpk_utils.xpk_print(f'List Job request returned ERROR {return_code}')
-    xpk_utils.xpk_exit(return_code)
+    xpk_print(f'List Job request returned ERROR {return_code}')
+    xpk_exit(return_code)
 
   lines = return_msg.split('\n')
   new_workload_name = args.workload
@@ -2535,7 +1964,7 @@ def check_if_workload_can_schedule(args, system: SystemCharacteristics) -> bool:
 
   # Prevents workload creation failure for existing clusters with no ConfigMap
   if cluster_config_map is None:
-    xpk_utils.xpk_print(
+    xpk_print(
         'No ConfigMap exist for cluster with the name'
         f' {resources_configmap_name}.'
     )
@@ -2544,7 +1973,7 @@ def check_if_workload_can_schedule(args, system: SystemCharacteristics) -> bool:
   # Check for gke accelerator type:
   missing_gke_accelerator_type = False
   if system.gke_accelerator not in cluster_config_map:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Gke Accelerator Type Check: {args.workload} is requesting'
         f' {system.gke_accelerator} but cluster only contains'
         f' {cluster_config_map.keys()}. '
@@ -2561,7 +1990,7 @@ def check_if_workload_can_schedule(args, system: SystemCharacteristics) -> bool:
     num_chips_in_workload = get_total_chips_requested_from_args(args, system)
 
     if num_chips_in_workload > max_chips_in_cluster:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'{args.workload} is requesting {num_chips_in_workload} chips but'
           f' the cluster {args.cluster} supports up to {max_chips_in_cluster}.'
           '  Resize the cluster to support more chips with'
@@ -2574,14 +2003,14 @@ def check_if_workload_can_schedule(args, system: SystemCharacteristics) -> bool:
   missing_device_type = False
   device_type = system.device_type
   if device_type not in cluster_config_map:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Device Type Check: {args.workload} is requesting {device_type} but '
         f'cluster only contains {cluster_config_map.keys()}. '
     )
     missing_device_type = True
 
   if missing_device_type and missing_gke_accelerator_type:
-    xpk_utils.xpk_print(
+    xpk_print(
         'Both Device Type and GKE Accelerator Type checks failed.'
         f' XPK will not create the workload {args.workload}.'
     )
@@ -2594,7 +2023,7 @@ def check_if_workload_can_schedule(args, system: SystemCharacteristics) -> bool:
     else:
       vm_required_by_workload = args.num_slices * system.vms_per_slice
     if vm_required_by_workload > max_vm_in_cluster:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'{args.workload} is requesting {args.num_slices} slice/slices of'
           f' {device_type}, which is {vm_required_by_workload} VMs, but the'
           f' cluster only contains {max_vm_in_cluster} VMs of {device_type}.'
@@ -2618,17 +2047,17 @@ def use_base_docker_image_or_docker_image(args) -> bool:
   # Check if (base_docker_image and script_dir) or (docker_image) is set.
   if args.docker_image is not None:
     if args.script_dir is not default_script_dir:
-      xpk_utils.xpk_print(
+      xpk_print(
           '`--script-dir` and --docker-image can not be used together. Please'
           ' see `--help` command for more details.'
       )
-      xpk_utils.xpk_exit(1)
+      xpk_exit(1)
     if args.base_docker_image is not default_docker_image:
-      xpk_utils.xpk_print(
+      xpk_print(
           '`--base-docker-image` and --docker-image can not be used together.'
           ' Please see `--help` command for more details.'
       )
-      xpk_utils.xpk_exit(1)
+      xpk_exit(1)
     use_base_docker_image = False
   return use_base_docker_image
 
@@ -2650,17 +2079,17 @@ def setup_docker_image(args) -> tuple[int, str]:
   if use_base_docker_image:
     validate_docker_image_code = validate_docker_image(docker_image, args)
     if validate_docker_image_code != 0:
-      xpk_utils.xpk_exit(validate_docker_image_code)
+      xpk_exit(validate_docker_image_code)
     build_docker_image_code, docker_image = build_docker_image_from_base_image(
         args
     )
     if build_docker_image_code != 0:
-      xpk_utils.xpk_exit(build_docker_image_code)
+      xpk_exit(build_docker_image_code)
   else:
     docker_image = args.docker_image
     validate_docker_image_code = validate_docker_image(args.docker_image, args)
     if validate_docker_image_code != 0:
-      xpk_utils.xpk_exit(validate_docker_image_code)
+      xpk_exit(validate_docker_image_code)
 
   return 0, docker_image
 
@@ -2746,7 +2175,7 @@ def get_main_container(args, system, docker_image, resource_type) -> str:
   xpk_return_user_exit_code = ''
   if args.restart_on_user_code_failure:
     if int(args.max_restarts) <= 0:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'Warning: --max-restarts, is set to {args.max_restarts}. Will not'
           ' restart on user failure.'
       )
@@ -2916,149 +2345,6 @@ def get_volume_mounts(args, system: SystemCharacteristics) -> str:
   return volume_mount_yaml
 
 
-def get_pathways_rm_args(args, system: SystemCharacteristics) -> str:
-  """Arguments for the Pathways resource manager.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: yaml containing arguments for the Pathways resource manager.
-  """
-  yaml = """- --alsologtostderr
-              - --pathways_server_port=38677
-              - --pathways_server_provides_devices=false
-              - --pathways_device_type=NONE
-              - --pathways_persistent_compilation_cache=false
-              - --pathways_tmp_dir_pattern={args.pathways_gcs_location}
-              - --pathways_expected_instances={expected_instances}"""
-  if args.use_pathways:
-    return yaml.format(
-        args=args,
-        expected_instances=compute_pathways_expected_instances(args, system),
-    )
-  else:
-    return ''
-
-
-def compute_pathways_expected_instances(
-    args, system: SystemCharacteristics
-) -> str:
-  """Computes the expected instances from the system characteristics.
-  Args:
-    args: user provided args.
-    system: system characteristics.
-
-  Returns:
-    str: formatted string representing the expected instances (eg:
-    "tpuv4:2x2x2,tpuv4:2x2x2" for 2 slices of v4-16).
-  """
-  expected_instances = ','.join([
-      f'tpu{get_pathways_expected_tpu_type(system.device_type)}:{system.topology}'
-      for _ in range(args.num_slices)
-  ])
-
-  xpk_utils.xpk_print(f'Pathways expected instances are: {expected_instances}')
-  return expected_instances
-
-
-def get_pathways_expected_tpu_type(device_type: str) -> str:
-  """Returns the device type expected by Pathways
-  Args:
-    device_type: the system characteristic device type
-
-  Returns:
-    str: the device type expected by pathways.
-  """
-  raw_type = device_type.split('-')[0].lower()
-  pathways_expected_instance = PathwaysExpectedInstancesMap[raw_type]
-  if not pathways_expected_instance:
-    xpk_utils.xpk_print(
-        f'Passed in device_type {device_type} is incorrect. Please pass in a'
-        ' valid device type'
-    )
-    xpk_utils.xpk_exit(1)
-  return pathways_expected_instance
-
-
-def get_rm_address(args) -> str:
-  """Generates the Pathways resource manager address based on whether CloudDNS is enabled or not.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: Fully qualified RM address.
-  """
-  suffix = ''
-  if is_cluster_using_clouddns(args):
-    suffix = f'.default.svc.{args.cluster}-domain.'
-  rm_address = f'{args.workload}-rm-0-0.{args.workload}{suffix}:38677'
-  return rm_address
-
-
-def get_proxy_address(args) -> str:
-  """Generates the Pathways proxy address based on whether CloudDNS is enabled or not.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: Fully qualified proxy address.
-  """
-  suffix = ''
-  if is_cluster_using_clouddns(args):
-    suffix = f'.default.svc.{args.cluster}-domain.'
-  proxy_address = (
-      f'grpc://{args.workload}-proxy-0-0.{args.workload}{suffix}:38676'
-  )
-  return proxy_address
-
-
-def get_pathways_worker_args(args) -> str:
-  """Arguments for the Pathways workers.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: yaml containing arguments for the Pathways workers.
-  """
-  yaml = """- --alsologtostderr
-              - --pathways_server_port=38677
-              - --pathways_resource_manager={rm_address}
-              - --pathways_persistent_compilation_cache=false
-              - --xla_tpu_enable_data_parallel_all_reduce_opt=true
-              - --xla_tpu_data_parallel_opt_different_sized_ops=true
-              - --xla_tpu_enable_async_collective_fusion=true
-              - --xla_tpu_enable_async_collective_fusion_fuse_all_gather=true
-              - --xla_tpu_enable_async_collective_fusion_multiple_steps=true
-              - --xla_tpu_overlap_compute_collective_tc=true
-              - --xla_enable_async_all_gather=true
-              - --pathways_tmp_dir_pattern={args.pathways_gcs_location}"""
-  if args.use_pathways:
-    return yaml.format(args=args, rm_address=get_rm_address(args))
-  else:
-    return ''
-
-
-def get_pathways_proxy_args(args) -> str:
-  """Arguments for the Pathways proxy.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: yaml containing arguments for the Pathways proxy.
-  """
-  yaml = """- --alsologtostderr
-              - --v=0
-              - --pathways_ifrt_proxy_server_resource_manager={rm_address}
-              - --pathways_ifrt_proxy_server_port=38676
-              - --pathways_tmp_dir_pattern={args.pathways_gcs_location}
-              - --pathways_plaque_network=gcp"""
-
-  if args.use_pathways:
-    return yaml.format(args=args, rm_address=get_rm_address(args))
-  else:
-    return ''
-
-
 def get_user_workload_container(args, system: SystemCharacteristics):
   """Deploy user workload container
 
@@ -3073,7 +2359,7 @@ def get_user_workload_container(args, system: SystemCharacteristics):
 
   setup_docker_image_code, docker_image = setup_docker_image(args)
   if setup_docker_image_code != 0:
-    xpk_utils.xpk_exit(setup_docker_image_code)
+    xpk_exit(setup_docker_image_code)
 
   # Determine if we deploy a sidecar and if we deploy a container.
   debugging_dashboard_id = None
@@ -3085,7 +2371,7 @@ def get_user_workload_container(args, system: SystemCharacteristics):
       and system.accelerator_type == AcceleratorType['TPU']
       and args.deploy_stacktrace_sidecar
   ):
-    xpk_utils.xpk_print(
+    xpk_print(
         'Sidecar container to display stack traces for TPU workloads will also'
         ' be deployed.'
     )
@@ -3095,49 +2381,6 @@ def get_user_workload_container(args, system: SystemCharacteristics):
   else:
     container = get_main_container(args, system, docker_image, resource_type)
   return container, debugging_dashboard_id
-
-
-def get_user_workload_for_pathways(args, system: SystemCharacteristics) -> str:
-  """
-  Create a user workload container for Pathways.
-  Don't create one for Pathways headless mode.
-
-  Args:
-    args: user provided args.
-    system: system characteristics.
-
-
-  Returns:
-    str:
-      Pathways server port as a YAML string
-  """
-  user_workload_yaml = """- name: main
-    replicas: 1
-    template:
-      metadata:
-        labels:
-          xpk.google.com/workload: {args.workload}
-      spec:
-        backoffLimit: 0
-        completions: 1
-        parallelism: 1
-        template:
-          spec:
-            containers:
-              {container}
-            nodeSelector:
-              cloud.google.com/gke-nodepool: cpu-user-np
-            restartPolicy: OnFailure
-            volumes:
-            - hostPath:
-                path: /tmp
-                type: DirectoryOrCreate
-              name: shared-tmp"""
-  if args.headless:
-    return ''
-  else:
-    container, _ = get_user_workload_container(args, system)
-    return user_workload_yaml.format(args=args, container=container)
 
 
 def get_env_container(args, system: SystemCharacteristics):
@@ -3297,7 +2540,7 @@ def get_gke_dashboard(args, dashboard_filter):
   )
 
   if return_code != 0:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'GKE Dashboard List request returned ERROR {return_code}. If there is'
         ' a permissions error, please check'
         ' https://github.com/google/xpk/blob/main/README.md#roles-needed-based-on-permission-errors'
@@ -3306,7 +2549,7 @@ def get_gke_dashboard(args, dashboard_filter):
     return True, None
 
   if not return_value:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'No dashboard with {dashboard_filter} found in the'
         f' project:{args.project}.'
     )
@@ -3314,7 +2557,7 @@ def get_gke_dashboard(args, dashboard_filter):
 
   dashboards = return_value.strip().split('\n')
   if len(dashboards) > 1:
-    xpk_utils.xpk_print(
+    xpk_print(
         f'Multiple dashboards with same {dashboard_filter} exist in the'
         f' project:{args.project}. Delete all but one dashboard deployed using'
         ' https://github.com/google/cloud-tpu-monitoring-debugging.'
@@ -3347,7 +2590,7 @@ def get_gke_outlier_dashboard(args):
 
   # 'gcloud monitoring dashboards list' succeeded but no dashboard for the filter exist in the project
   if not is_error and not dashboard_id:
-    xpk_utils.xpk_print(
+    xpk_print(
         'Follow https://github.com/google/cloud-tpu-monitoring-debugging to'
         ' deploy monitoring dashboard to view statistics and outlier mode of'
         ' GKE metrics.'
@@ -3377,7 +2620,7 @@ def get_gke_debugging_dashboard(args):
 
   # 'gcloud monitoring dashboards list' succeeded but no dashboard for the filter exist in the project
   if not is_error and not dashboard_id:
-    xpk_utils.xpk_print(
+    xpk_print(
         'Follow https://github.com/google/cloud-tpu-monitoring-debugging to'
         ' deploy debugging dashboard to view stack traces collected in Cloud'
         ' Logging.'
@@ -3505,23 +2748,6 @@ def get_cpu_affinity(accelerator_type) -> str:
   return ''
 
 
-def get_pathways_unified_query_link(args) -> str:
-  """Get the unified query link for the pathways workload."""
-  pw_suffixes = ['main', 'rm', 'proxy']
-  pw_pod_names = [f'"{args.workload}-{suffix}-0"' for suffix in pw_suffixes]
-  pw_pod_names_query = '%20OR%20'.join(pw_pod_names + ['worker-0-0'])
-  query_params = (
-      'resource.type%3D"k8s_container"%0A'
-      f'resource.labels.project_id%3D"{args.project}"%0A'
-      f'resource.labels.location%3D"{zone_to_region(args.zone)}"%0A'
-      f'resource.labels.cluster_name%3D"{args.cluster}"%0A'
-      f'resource.labels.pod_name:{pw_pod_names_query}%0A'
-      'severity>%3DDEFAULT'
-  )
-
-  return f'https://console.cloud.google.com/logs/query;query={query_params}'
-
-
 def get_gpu_scheduler(
     args, system: SystemCharacteristics, autoprovisioning_args: str
 ) -> tuple[str, int]:
@@ -3555,7 +2781,7 @@ def get_gpu_scheduler(
     )
   else:
     return_code = 1
-    xpk_utils.xpk_print(
+    xpk_print(
         '--scheduler needs to be set as either `default-scheduler`'
         ' or `gke.io/topology-aware-auto` in order to schedule the'
         ' workloads on GPUs.'
@@ -3661,171 +2887,6 @@ def get_gpu_tcp_volume(system: SystemCharacteristics) -> str:
   return gpu_tcp_volume
 
 
-def ensure_pathways_workload_prerequisites(args, system) -> bool:
-  """Check all Pathways workload prerequisites and set necessary args.
-
-  Args:
-    args: user provided arguments for running the command.
-    system: system characteristics.
-
-  Returns:
-    True once conditions satisfy and variables are set. Exits otherwise.
-  """
-  # Ensure command is provided if not using Pathways in headless mode
-  if args.command is None and not args.headless:
-    xpk_utils.xpk_print(
-        'Please provide a command using "--command" for the docker container to'
-        ' execute. Command is not required if you wish to run Pathways'
-        ' workloads in headless mode (`xpk workload create-pathways'
-        ' --headless`).'
-    )
-    xpk_utils.xpk_exit(1)
-
-  # Ensure the cluster and CPU nodepools were created with create-pathways
-  all_node_pools = get_all_nodepools_programmatic(args)
-  desired_pw_cpu_node_pools = {'cpu-user-np', 'cpu-rm-np', 'cpu-proxy-np'}
-  if not desired_pw_cpu_node_pools.issubset(set(all_node_pools[0])):
-    xpk_utils.xpk_print(
-        'Cluster needs to be created with `xpk create-pathways` to run'
-        ' Pathways workloads.'
-    )
-    xpk_utils.xpk_exit(1)
-
-  # Ensure device type is TPUs - currently Pathways supports TPUs only.
-  if system.accelerator_type != AcceleratorType['TPU']:
-    xpk_utils.xpk_print(
-        'Currently, Pathways workloads can only be run on TPUs.'
-    )
-    xpk_utils.xpk_exit(1)
-
-  # Set proxy address to be consumed in helper methods and displayed to user.
-  args.pathways_proxy_address = get_proxy_address(args)
-
-  # Set the job which determines the life of other Pathways jobs
-  args.targetReplicatedJob = 'proxy' if args.headless else 'main'
-
-  # Always report user code failures back to JobSet.
-  args.restart_on_user_code_failure = True
-
-  return True
-
-
-def workload_list_awk_command(filter_key) -> str:
-  """Function returns the awk command needed from the filter specified.
-
-  Args:
-    filter_key: workload list filter to awk against
-
-  Returns:
-    awk command to use in filtering workload list.
-  """
-
-  return f" | awk -e 'NR == 1 || {filter_key} {{print $0}}'"
-
-
-def determine_workload_list_filter_by_status(args) -> str:
-  """Function to create the filtered view of workload list.
-
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    the argument needed to filter by status of jobs in workload list.
-  """
-  # Argument positions related to columns created by workload list command.
-  status_arg = '$7'
-  running_vms_arg = '$5'
-  status_verbose_arg = '$9'
-  if args.filter_by_status == 'EVERYTHING':
-    return ''
-  elif args.filter_by_status == 'RUNNING':
-    # Running includes the status Admitted or Evicted, and when the number of
-    # vms running is > 0.
-    return workload_list_awk_command(
-        f'({status_arg} ~ "Admitted|Evicted" && {running_vms_arg} ~ /^[0-9]+$/'
-        f' && {running_vms_arg} > 0)'
-    )
-  elif args.filter_by_status == 'QUEUED':
-    # Queued includes the status Admitted or Evicted, and when the number of
-    # vms running is 0.
-    return workload_list_awk_command(
-        f'({status_arg} ~ "Admitted|Evicted|QuotaReserved" &&'
-        f' ({running_vms_arg} ~ "<none>" || {running_vms_arg} == 0))'
-    )
-  elif args.filter_by_status == 'FINISHED':
-    return workload_list_awk_command(f'{status_arg} == "Finished"')
-  elif args.filter_by_status == 'FAILED':
-    # Failed includes the status Finished, and when the verbose reason is failed.
-    return workload_list_awk_command(
-        f'({status_arg} == "Finished" && {status_verbose_arg} ~ "failed")'
-    )
-  elif args.filter_by_status == 'SUCCESSFUL':
-    # Failed includes the status Finished, and when the verbose reason is finished/success.
-    return workload_list_awk_command(
-        f'({status_arg} == "Finished" && {status_verbose_arg} ~ "finished")'
-    )
-  raise RuntimeError(f'Can not find filter type: {args.filter_by_status}')
-
-
-def determine_workload_list_filter_by_job(args) -> str:
-  """Function to filter view of workload list based on job name.
-
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    the argument needed to filter job names from workload list
-  """
-  # Argument positions related to columns created by workload list command.
-  if not args.filter_by_job:
-    return ''
-  else:
-    job_name_arg = '$1'
-    return workload_list_awk_command(f'{job_name_arg} ~ "{args.filter_by_job}"')
-
-
-def get_workload_list(args) -> tuple[int, str]:
-  """Function to get the list of the workloads in the cluster.
-
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    return_code: 0 if successful and 1 otherwise.
-    return_value: workloads in the cluster matching the criteria.
-  """
-  columns = {
-      'Jobset Name': '.metadata.ownerReferences[0].name',
-      'Created Time': '.metadata.creationTimestamp',
-      'Priority': '.spec.priorityClassName',
-      'TPU VMs Needed': '.spec.podSets[0].count',
-      'TPU VMs Running/Ran': '.status.admission.podSetAssignments[-1].count',
-      'TPU VMs Done': '.status.reclaimablePods[0].count',
-      'Status': '.status.conditions[-1].type',
-      'Status Message': '.status.conditions[-1].message',
-      'Status Time': '.status.conditions[-1].lastTransitionTime',
-  }
-  s = ','.join([key + ':' + value for key, value in columns.items()])
-
-  workload_list_filter_status_cmd = determine_workload_list_filter_by_status(
-      args
-  )
-  workload_list_filter_job_cmd = determine_workload_list_filter_by_job(args)
-  command = (
-      f'kubectl get workloads -o=custom-columns="{s}" '
-      f'{workload_list_filter_status_cmd} {workload_list_filter_job_cmd}'
-  )
-
-  return_code, return_value = run_command_for_value(
-      command,
-      f'List Jobs with filter-by-status={args.filter_by_status}'
-      f' with filter-by-job={args.filter_by_job}',
-      args,
-  )
-
-  return return_code, return_value
-
-
 def wait_for_job_completion(args) -> int:
   """Function to wait for job completion.
 
@@ -3839,7 +2900,7 @@ def wait_for_job_completion(args) -> int:
   args.workload = args.wait_for_job_completion
   workload_exists = check_if_workload_exists(args)
   if not workload_exists:
-    xpk_utils.xpk_print(f'Workload named {args.workload} does not exist.')
+    xpk_print(f'Workload named {args.workload} does not exist.')
     return 1
 
   # Get the full workload name
@@ -3848,9 +2909,7 @@ def wait_for_job_completion(args) -> int:
       get_workload_name_cmd, 'Get full workload name', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
-        f'Get full workload name request returned ERROR {return_code}'
-    )
+    xpk_print(f'Get full workload name request returned ERROR {return_code}')
     return return_code
   full_workload_name = return_value.split(' ')[0]
 
@@ -3871,7 +2930,7 @@ def wait_for_job_completion(args) -> int:
   )
   if return_code != 0:
     if 'timed out' in return_value:
-      xpk_utils.xpk_print(
+      xpk_print(
           f'Timed out waiting for your workload after {timeout_msg}, see your'
           ' workload here:'
           # pylint: disable=line-too-long
@@ -3879,10 +2938,10 @@ def wait_for_job_completion(args) -> int:
       )
       return 124
     else:
-      xpk_utils.xpk_print(f'{return_value}')
-      xpk_utils.xpk_print(f'Wait for workload returned ERROR {return_code}')
+      xpk_print(f'{return_value}')
+      xpk_print(f'Wait for workload returned ERROR {return_code}')
       return return_code
-  xpk_utils.xpk_print(
+  xpk_print(
       'Finished waiting for your workload, see your workload here:'
       # pylint: disable=line-too-long
       f' https://console.cloud.google.com/kubernetes/service/{zone_to_region(args.zone)}/{args.cluster}/default/{args.workload}/details?project={args.project}'
@@ -3895,12 +2954,10 @@ def wait_for_job_completion(args) -> int:
       status_cmd, 'Get jobset status', args
   )
   if return_code != 0:
-    xpk_utils.xpk_print(
-        f'Get workload status request returned ERROR {return_code}'
-    )
+    xpk_print(f'Get workload status request returned ERROR {return_code}')
     return return_code
-  xpk_utils.xpk_print(f'Your workload finished with status: {return_value}')
+  xpk_print(f'Your workload finished with status: {return_value}')
   if return_value != 'Completed':
-    xpk_utils.xpk_print('Your workload did not complete successfully')
+    xpk_print('Your workload did not complete successfully')
     return 125
   return 0
